feat: roughly break into components.
